refactor: #3 마트 섹션 리팩토링
